@@ -1,14 +1,10 @@
 ﻿using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using MTWireGuard.Application.Utils;
 
 namespace MTWireGuard.Application
 {
-    public class ApplicationLifetimeService(IHostApplicationLifetime hostApplicationLifetime, ILogger logger) : IHostedService
+    public class ApplicationLifetimeService(IHostApplicationLifetime hostApplicationLifetime, ILogger<ApplicationLifetimeService> logger) : IHostedService
     {
         public Task StartAsync(CancellationToken cancellationToken)
         {
@@ -24,17 +20,21 @@ namespace MTWireGuard.Application
 
         private void OnStarted()
         {
-            logger.LogInformation("Registered: {Name}!", Environment.UserName);
+            logger.LogInformation("Application started successfully - User: {UserName}, Machine: {MachineName}", 
+                Environment.UserName, Environment.MachineName);
         }
 
         private void OnStopping()
         {
-            logger.LogWarning("Exiting: {MachineName}!", Environment.MachineName);
+            Console.ForegroundColor = ConsoleColor.Yellow;
+            Console.WriteLine("Shutting down...");
+            Console.ResetColor();
+            logger.LogWarning("Application is stopping - Machine: {MachineName}", Environment.MachineName);
         }
 
         private void OnStopped()
         {
-            // ...
+            logger.LogInformation("Application stopped");
         }
     }
 }
