@@ -146,7 +146,6 @@
             <div class="card-body text-center">
                 <img class="w-50" src="img/logo.png" alt="" srcset="">
                 <small class="d-block mb-1">v@(Application.Utils.CoreUtil.GetProjectVersion())</small>
-                <small class="d-block mb-1">ID: @(Application.Utils.CoreUtil.GetIDContent())</small>
                 <ul class="list-group list-unstyled border border-2 rounded-3">
                     <li>
                         <a class="list-group-item d-flex align-items-center bg-gradient-green" href="https://MTWireguard.techgarage.ir/">
@@ -161,7 +160,7 @@
                         </a>
                     </li>
                     <li>
-                        <a class="list-group-item d-flex align-items-center bg-gradient-red" href="https://github.com/techgarage-ir/MTWireGuard/">
+                        <a class="list-group-item d-flex align-items-center bg-gradient-red" href="https://github.com/dot-mike/MTWireGuard/">
                             <i class='bx bxl-github ms-2 me-4'></i>
                             <span>Github</span>
                         </a>
