@@ -643,7 +643,7 @@ $(function () {
             return;
         }).finally(() => {
             dt_basic.ajax.reload();
-            new bootstrap.Modal('#RemoveModal').hide();
+            bootstrap.Modal.getInstance(document.getElementById('RemoveModal')).hide();
         });
     });
 
@@ -663,7 +663,7 @@ $(function () {
                     return;
                 }).finally(() => {
                     dt_basic.ajax.reload();
-                    new bootstrap.Modal('#ImportModal').hide();
+                    bootstrap.Modal.getInstance(document.getElementById('ImportModal')).hide();
                 });
             } catch (error) {
                 console.error('Invalid JSON:', error);
